My Application packaged:
Packaging my application along with its binaries, configurations, run-time environment, dependencies.
A Dockerfile is included which includes instructions to choose a base image as a production environment, deploy the entire source code and run and test the application in the production environment.
Auto-build process in Docker hub:
docker-compose.test.yml file is added that runs the run_tests.sh test cases in the container environment.